Consolidate to_char
