+++ datafusion-examples/examples/sql.rs
@@ -19,8 +19,11 @@
 //!
 //! * [`parquet`]: run SQL query against a single Parquet file
 //! * [`parquet_multi_files`]: run SQL query against a table backed by multiple Parquet files
-//! * [`regexp`]: regular expression functions
+//! * [`regexp`]: regular expression functions to manipulate strings
+//! * [`to_char`]: to_char function to convert strings to date, time, timestamp and durations
 
+use arrow::array::{Date32Array, RecordBatch, StringArray};
+use arrow_schema::{DataType, Field, Schema};
 use datafusion::datasource::file_format::parquet::ParquetFormat;
 use datafusion::datasource::listing::ListingOptions;
 use datafusion::error::Result;
@@ -35,6 +38,7 @@ async fn main() -> Result<()> {
     parquet().await?;
     parquet_multi_files().await?;
     regexp().await?;
+    to_char().await?;
     Ok(())
 }
 
@@ -471,6 +475,196 @@ async fn regexp() -> Result<()> {
             "+---------------------------------------------------------------+",
             "|                                                               |",
             "+---------------------------------------------------------------+",
+        ],
+        &result
+    );
+
+    Ok(())
+}
+
+/// This example demonstrates how to use the to_char function via sql
+///
+/// This function accepts date, time, timestamp and duration values
+/// in the first argument and string values for the second
+async fn to_char() -> Result<()> {
+    let schema = Arc::new(Schema::new(vec![
+        Field::new("values", DataType::Date32, false),
+        Field::new("patterns", DataType::Utf8, false),
+    ]));
+
+    let batch = RecordBatch::try_new(
+        schema,
+        vec![
+            Arc::new(Date32Array::from(vec![18506, 18507, 18508, 18509])),
+            Arc::new(StringArray::from(vec![
+                "%Y-%m-%d", "%Y:%m:%d", "%Y%m%d", "%d-%m-%Y",
+            ])),
+        ],
+    )?;
+
+    // declare a new context. In spark API, this corresponds to a new spark SQLsession
+    let ctx = SessionContext::new();
+
+    // declare a table in memory. In spark API, this corresponds to createDataFrame(...).
+    ctx.register_batch("t", batch)?;
+    let _ = ctx.table("t").await?;
+
+    // use to_char function to convert col 'values' to timestamp type using
+    // patterns stored in col 'patterns'
+    let result = ctx
+        .sql("SELECT to_char(values, patterns) from t")
+        .await?
+        .collect()
+        .await?;
+
+    assert_batches_eq!(
+        &[
+            "+------------------------------+",
+            "| to_char(t.values,t.patterns) |",
+            "+------------------------------+",
+            "| 2020-09-01                   |",
+            "| 2020:09:02                   |",
+            "| 20200903                     |",
+            "| 04-09-2020                   |",
+            "+------------------------------+",
+        ],
+        &result
+    );
+
+    // the date_format alias for the to_char function can be used as well
+    let result = ctx
+        .sql("SELECT date_format(values, patterns) from t")
+        .await?
+        .collect()
+        .await?;
+
+    assert_batches_eq!(
+        &[
+            "+------------------------------+",
+            "| to_char(t.values,t.patterns) |",
+            "+------------------------------+",
+            "| 2020-09-01                   |",
+            "| 2020:09:02                   |",
+            "| 20200903                     |",
+            "| 04-09-2020                   |",
+            "+------------------------------+",
+        ],
+        &result
+    );
+
+    // use to_char function to convert col 'values' with a fixed format
+    let result = ctx
+        .sql("SELECT to_char(values, '%m-%d-%Y') FROM t")
+        .await?
+        .collect()
+        .await?;
+
+    assert_batches_eq!(
+        &[
+            "+------------------------------------+",
+            "| to_char(t.values,Utf8(\"%m-%d-%Y\")) |",
+            "+------------------------------------+",
+            "| 09-01-2020                         |",
+            "| 09-02-2020                         |",
+            "| 09-03-2020                         |",
+            "| 09-04-2020                         |",
+            "+------------------------------------+",
+        ],
+        &result
+    );
+
+    // if you want to just use the default format cast to a string
+    let result = ctx
+        .sql("SELECT arrow_cast(values, 'Utf8') from t")
+        .await?
+        .collect()
+        .await?;
+
+    assert_batches_eq!(
+        &[
+            "+-----------------------------------+",
+            "| arrow_cast(t.values,Utf8(\"Utf8\")) |",
+            "+-----------------------------------+",
+            "| 2020-09-01                        |",
+            "| 2020-09-02                        |",
+            "| 2020-09-03                        |",
+            "| 2020-09-04                        |",
+            "+-----------------------------------+",
+        ],
+        &result
+    );
+
+    // use can use literals as well (note the use of timestamp here)
+    let result = ctx
+        .sql("SELECT to_char(arrow_cast(TIMESTAMP '2023-08-03 14:38:50Z', 'Timestamp(Second, None)'), '%d-%m-%Y %H:%M:%S')")
+        .await?
+        .collect()
+        .await?;
+
+    assert_batches_eq!(
+        &[
+            "+-------------------------------------------------------------------------------------------------------------+",
+            "| to_char(arrow_cast(Utf8(\"2023-08-03 14:38:50Z\"),Utf8(\"Timestamp(Second, None)\")),Utf8(\"%d-%m-%Y %H:%M:%S\")) |",
+            "+-------------------------------------------------------------------------------------------------------------+",
+            "| 03-08-2023 14:38:50                                                                                         |",
+            "+-------------------------------------------------------------------------------------------------------------+",
+        ],
+        &result
+    );
+
+    // durations are supported though the output format is limited to two formats
+    // 'pretty' and 'ISO8601'
+    let result = ctx
+        .sql("SELECT to_char(arrow_cast(123456, 'Duration(Second)'), 'pretty')")
+        .await?
+        .collect()
+        .await?;
+
+    assert_batches_eq!(
+        &[
+            "+----------------------------------------------------------------------------+",
+            "| to_char(arrow_cast(Int64(123456),Utf8(\"Duration(Second)\")),Utf8(\"pretty\")) |",
+            "+----------------------------------------------------------------------------+",
+            "| 1 days 10 hours 17 mins 36 secs                                            |",
+            "+----------------------------------------------------------------------------+",
+        ],
+        &result
+    );
+
+    // durations are supported though the output format is limited to two formats
+    // 'pretty' and 'ISO8601'
+    let result = ctx
+        .sql("SELECT to_char(arrow_cast(123456, 'Duration(Second)'), 'iso8601')")
+        .await?
+        .collect()
+        .await?;
+
+    assert_batches_eq!(
+        &[
+            "+-----------------------------------------------------------------------------+",
+            "| to_char(arrow_cast(Int64(123456),Utf8(\"Duration(Second)\")),Utf8(\"iso8601\")) |",
+            "+-----------------------------------------------------------------------------+",
+            "| PT123456S                                                                   |",
+            "+-----------------------------------------------------------------------------+",
+        ],
+        &result
+    );
+
+    // output format is null
+
+    let result = ctx
+        .sql("SELECT to_char(arrow_cast(123456, 'Duration(Second)'), null) as result")
+        .await?
+        .collect()
+        .await?;
+
+    assert_batches_eq!(
+        &[
+            "+--------+",
+            "| result |",
+            "+--------+",
+            "|        |",
+            "+--------+",
         ],
         &result
     );
